+++ CHANGELOG.md
@@ -6,8 +6,17 @@
 * COMPILER ASSISTED BREAKING CHANGE: Not exporting `GFlay'`, `gflay'`,
   `gterminal` anymore.
 
-* COMPILER ASSISTED BREAKING CHANGE: `GFlay1` and `All` are now type synonyms
+* COMPILER ASSISTED BREAKING CHANGE: `GFlay` and `All` are now type synonyms
   rather than classes.
+
+* COMPILER ASSISTED BREAKING CHANGE: `Flayable1` is not a typeclass anymore.
+  Now it is a type synonym for `Flayable` using “quantified constraints” as
+  implemented in a spirit similar to `Data.Constraint.Forall`.
+  Accordingly, `flay1` is not a typeclass method anymore.
+
+* `Generic` types parametrized by a type constructor who have all of their
+  immediate children fields using said type constructor get a `Flayable`
+  instance for free.
 
 * Added `pump`, `dump`, `Pump` and `GPump`.